refactor(transport layer): define interfaces in consumer packages
